refactor: rename SplitCalendarButton to CalendarSplitButton for clarity and backward compatibility; introduce CalendarButton and CalendarCard components; update stories accordingly
