+++ docs/tasks/01 - Calendar Atoms.md
@@ -1,0 +1,5 @@
+
+ lets now look at the other options of how calendar are use int he app.. and resturcture to match these atoms for the system to use.
+  - CalendarCard
+  - CalendarButton - this is a single button text option with a calenader icon that creates a popout calender display
+  - CalendarSplitButton - is is what the current SplitCalendarButton is, just need to rename it.
